adaptive_dsa_agent/requirements.txt
# Runtime dependencies for the Adaptive DSA Tutor Agent.
# The Gemini dependency is OPTIONAL — the hint generator auto-falls-back to an
# offline rule-based hinter if the package or API key is unavailable.

google-generativeai>=0.8.0 ; python_version >= "3.10"

# HTTP API + database (FastAPI server in ``server/``)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
google-auth>=2.29.0
bcrypt>=4.1.0
